+++ activerecord/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Set polymorphic type column to NULL on `dependent: :nullify` strategy. 
+    
+    On polymorphic associations both the foreign key and the foreign type columns will be set to NULL.
+    
+    *Laerti Papa*
 
 *   Allow `ActionController::Params` as argument of `ActiveRecord::Base#exists?`.